+++ weldx/utility.py
@@ -1,6 +1,7 @@
 """Contains package internal utility functions."""
 
 import math
+from collections.abc import Iterable
 from typing import Any, Dict, List, Union
 
 import numpy as np
@@ -479,7 +480,9 @@ def xr_interp_like(
     if isinstance(da2, (xr.DataArray, xr.Dataset)):
         sel_coords = da2.coords  # remember original interpolation coordinates
     else:  # assume da2 to be dict-like
-        sel_coords = da2
+        sel_coords = {
+            k: (v if isinstance(v, Iterable) else [v]) for k, v in da2.items()
+        }
 
     # store and strip pint units at this point, since the unit is lost during
     # interpolation and because of some other conflicts. Unit is restored before